Create scheduler for Node eviction selection logic
It makes a couple of simplifying assumptions to avoid replicating the
full logic of possible placements of VMs. This should eventually moved
to the scheduling service to avoid replicating the logic in different
places.

1. We only keep track of the resource changes done by the
controller itself. The issue is, that we otherwise would need to poll
the hypervisors. But that again is better solved directly in the
scheduler, as it has always up-to-date view of the resource usages.

2. We assume we can move VMs arbitrarily between the same hardware
class. A hardware-class is the CPU generation and the NUMA topology.

On one hand, that is too strict, as we can usually move VMs from an
older hardware class to a newer one, but not vice-versa.
Same with the NUMA topology of VMs, we can move a single NUMA Node
VM to a two-NUMA Node Host, but not a two NUMA Node VM to a single
NUMA Node.

To make things even more complicated, if we update the firmware,
or some software settings, we may not even be able to move VMs within
the same hardware class.

But due to the complexity of interactions of traits, and the need
to know the individual VMs architectures, that is again best solved
once in one place, and outside of this controller.