Current State for Review (#80)
* Decommissioning of hypervisor nodes

If labeled appropriately, the deletion of the node will be delayed
until the host is empty and has been removed from nova as a service.

* Initial Onboarding controller

Takes care of
- Creating certificates for the node
- Creating and adding the host to two aggregates
  1. The availability-zone aggregate matching the kubernetes failover-zone
  2. A (now hardcoded) test aggregate, which only allows specific projects
     to deploy to the hypervisor for testing purposes.
  3. Enablement of the service, so it can be tested
  4. Set a label on the node, that the node has been onboarded

It also sets two labels indicating the hypervisor and compute service id
in nova.

* Do not block deletion of eviction without hypervisor

If we decomission a node, we want to first evict it and then delete
it and never re-activate it, so it is a standard behaviour.

* Onboarding: Handle missing label gracefully, minor refactor

* EvictionLabel: Rely on the approriate label for the hostname

The hostname for the hypervisor is kubernetes.io/hostname.
The host for nova compute was used before.

* Consolidate some functions around labels

* Add eviction to decommission and unset labels to take out node

* Bump chart version

* Fix rbacs

* Temporarily disable certificates

* Ignore build charts

* Need to update nodes as well

* MaintenanceController replacement

Acts only on nodes which have a certain label set.
- Creates unevictable pods on tied to a node to stop gardener destroying the node
- Translates the termination event to a label as per maintenance controller
- Allows eviction of pod when maintenance has been allowed.

* Normalize constants

* More maintenance controller fixups

* Simplify decomission

* Keep node-labels

* Safeguard node deletion against VMs still on the host

* Added traits controller syncing annotations

* Maintenance: Label resource with metal-name, and scale down deployment

* EvictionLabel: Label with metal-name and addNodeOwner correctly

* Relabel hypervisor and ml2 node when allowing eviction

* Default to non-blocking, non-controller references

* Remove redundant  addNodeOwnerReference

SetControllerReference does the same plus sets controller true.

* Do not add to test aggregate for now

* disabling the tenant_filter_tests aggregate for now until work continues on this feature

* Fix verify resize

We first have to check, if the migration is still ongoing,
as the host changes before the status changes.

The result is, that we are not in verify_resize, but still in resize
for cold migrations, and therefor miss to verify it.

* Masakari: Onboarding/Offboarding and disable / enablement works

* Enable reauth

The operator is long-running, so the token will expire.
Instead of sprinkling the code with error handlers, we
can leave that to the reauth functionality in gophercloud.

* In setup only create one context

Less copy-and-paste.

* Fixup Enable reauth

* Put blocking deployments in HostNetwork mode

The agents are running in HostNetwork mode, and therefore
might already run even if there is no working pod networking.
Since they are running and might create VMs, we need to block
the teardown of the hosts in these cases as well.

* Reflect aggregate annotations to nova

This allows us to assign worker groups to certain
aggregates.

* Make masakari optional and do some smoke test

* Opt-out of testing

* Reflect onboarding test to gardener

The blocking pods are renamed to maint (for maintenance) and following the
documentation of node readiness in [gardener](https://gardener.cloud/docs/gardener/advanced/node-readiness/)
moved to the kube-system namespace and reflecting the state of the tests
in their startup probes.

* Skip eviction creation, if never onboarded

* Write a reason in the maintenance label for better visiblity

* Go mod tidy

* Split out certificate controller and activate

* Newlines thanks to new helmify

* Drop masakari support

* Drop lookup of name through label

* RetryOnConfict in maintenance-controller

* RetryOnConfict in node-certificate-controller

* Eviction Controller: Default to failure, if node cannot be found.

Not sure, if that is the right default, but it is the
one the tests ensure.

* NodeEvictionLabel: Needs onboarding label to trigger controller

* Workaround for tests: Set node kind and version

* NodeEvictionLabel: Pass less arguments to reconcileEviction

The arguments came from automatic refactoring, but are redundant.

* filtering status-only changes in the eviction_controller by adding a GenerationChangedPredicate

* Drop masakari

* Remove further masakari references

* removing GenerationChangedPredicate for now as it is too broad

* Onboarding: Scope the VM on the UID

We have now stable hostnames, so if we are stuck
with one VM for that host because some DB foo,
then we never will be able to continue.
To get it unstuck, use the UID. So a new node object
results in a new server.

* Typo fix

* Remove unused function

* Handle error in aggregates controller

* For loop is over keys by default

* Remove unused variables

* Remove unused return value

* Return errors, remove unused return value

* Removed embedded field selector

* More linter corrections

* Replaced addNodeOwnerReference with SetOwnerReference

* Removed unused rbacs, consts, and disabled code

* Update EvictionController RBAC for reading nodes

* Replaced hasAnyLabel with labels.Set...Has

* Use removeFinalizer in DecommissionController

* Clean password from non-printable chars

In case an accidental newline slips in somewhere.

* Volume type is now kvm-pilot

* NodeCertificateController: Increase certificate duration to 47 days

This matches the limit for CA/Browser Forum. It doesn't have to, but the reasoning
behind it probably applies to us to some degree as well.

* Revert "Clean password from non-printable chars"

This reverts commit 990ee190d55e6714f0f5935e6c65c2eb720274bc.

* Use sentinels instead of error types

---------

Co-authored-by: Marcus Christian Lehmann <marcus.christian.lehmann@sap.com>