decomission: Query servers directly, fix resource provider gone
The numServers on the hypervisor is synced asynchronously and
may be out of date.
So, rather query the compute api directly and check if there
is any server.

This should fix that the server gets removed from an aggregate
before all VMs have been migrated off the host.

Also, handle gracefully that the resource provider in placement
might have been removed concurrently.