Maintenance: Ensure there is always a blocking pod
We observed that during the node being terminated, that the pod
shortly disappeared due to some taint, which then allowed gardener
to tear the host down despite VMs being still active.

To counter that, tolerate any taint and effect.
Also change the deployment to a rolling-update,
so that instead of deleting that blocking pod and allowing
a teardown until the pod has been re-created, create it
first, and then let the other being torn down.