Switch Hypervisor from ControllerReference to OwnerReference
ControllerReference is 'for reconciling the owner object on changes to
controlled (with a Watch + EnqueueRequestForOwner)'.

Since we do not want to reconcile the node on changes to the hypervisor,
we are fine with OwnerReference.

The WithBlockOwnerDeletion ensures that the node is only
deleted, when the hypervisor resource also has been deleted.
For now, it is only avoiding a theoretical race conditition of the
node coming back before the hypervisor has been removed,
and annotations / labels not getting propagated.

It also sets the stage for moving the finalizer out of the node,
and doing that on the hypervisor resource instead.