Onboarding: Undo forceDown
An HA event on a node will cause the HA service to force down the node.
This is apparently persisted, so when the node comes up again, it stays
forced down, even though the service was deleted.

Either way, we can only onboard a node if it isn't `forcedDown`, so we
might as well ensure that.

We need to duplicate the gophercloud code until v3 is released, as the
bugfix is unfortunately a breaking api change.
See: https://github.com/gophercloud/gophercloud/pull/3531

In compliance with the Apache License, I kept the attribution and wrote
down where the code came from and what has been changed.