HypervisorMaintenanceController: Enable/Disable compute service
The controller only gets active after onboarding, as that one needs
to take care of enabling and aggregate association depending on the
tests.

If maintenance is set, it will disable now the compute host in nova,
and enable it, if it is unset.
It will only do so on an "edge", i.e. if it hasn't done it before.