Traits: Only reconcile after onboarding and before decomissioning
As with 6bb6c6 for the aggregates:

Traits set by the operator may conflict with the testing,
and decomissioning may make the resource-provider disappear.
To avoid these controllers fighting, let's limit this controller
to the actual active lifetime of the hypervisor.

Unfortunately, that means we have to reconcile on every status update.