Provide top-level visibility of kubectl modifications of hypervisor CROs
As we want also human operators to modify the CRO, at the very least to
set a hypervisor into maintenance, we need top-level visiblity of such
an action.

The HypervisorTaintController will check the managedFields for any
modification through kubeclt (kubectl-apply, kubectl-edit...) and add
a condition, which is prominently displayed in the overview table.