Refactor conditions, introduce terminating maintenance
Consolidate all condition to their respective API spec, this makes it
easier and clearer for external componenets to identify condition and
reason, also reuse conditions that are similar (Complated ->
Successful).

Also introducing a new maintenance enum value "terminating", that is
set by the maintenance-operator in case the node is terminating -
removing the need to monitor nodes status.